rotating mailjet key and upgrating the app
