+++ functions/lib/documents/documents.js
@@ -22,9 +22,6 @@ var __importStar = (this && this.__importStar) || function (mod) {
     __setModuleDefault(result, mod);
     return result;
 };
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.generatePDF = void 0;
 const storage_1 = require("firebase-admin/storage");
@@ -32,10 +29,10 @@ const functions = __importStar(require("firebase-functions"));
 const pdf_documents_1 = require("pdf-documents");
 const uuid = __importStar(require("uuid"));
 const business_1 = require("../business");
+const email_1 = require("../email/email");
 const index_1 = require("../index");
 const formats_1 = require("../util/formats");
 const whatsapp_1 = require("../whatsapp");
-const logger_1 = __importDefault(require("../util/logger"));
 const BUCKET_NAME = 'mobile-seller-documents';
 const LINK_DAYS_SIGNED = 604800;
 const sendWhatsappNotification = async (data, url, businessId) => {
@@ -98,7 +95,6 @@ const sendWhatsappNotification = async (data, url, businessId) => {
         const result = await (0, whatsapp_1.sendMessage)(template, currentToken, currentPhoneNumberId);
         if (result.status === 200) {
             functions.logger.info('Notification sent!', data.customer.name);
-            functions.logger.info('Notification sent! ' + data.customer.name);
         }
         else {
             functions.logger.error('Whatsapp notification could not be sent', result);
@@ -106,7 +102,6 @@ const sendWhatsappNotification = async (data, url, businessId) => {
     }
     catch (err) {
         functions.logger.error('Whatsapp notification could not be sent', err);
-        logger_1.default.error(err.message);
         throw new functions.https.HttpsError('cancelled', err.message);
     }
 };
@@ -114,7 +109,7 @@ const sendWhatsappNotification = async (data, url, businessId) => {
  * based on the user request it get the user who is requesting and get the business id associated
  */
 exports.generatePDF = functions.region(index_1.REGION).https.onCall(async (payload, context) => {
-    var _a, _b;
+    var _a, _b, _c, _d;
     try {
         functions.logger.info(payload);
         const requestedUser = await (0, index_1.getCurrentUserInfo)(context);
@@ -124,7 +119,7 @@ exports.generatePDF = functions.region(index_1.REGION).https.onCall(async (paylo
         const data = ['order', 'invoice', 'quote'].includes(payload.documentType)
             ? payload
             : payload;
-        console.info('data', data.customer.name);
+        console.info('Trying to send pdf document to: ', data.customer.name);
         const date = new Date();
         const day = date.getDate().toString().padStart(2, '0');
         const month = (date.getMonth() + 1).toString().padStart(2, '0');
@@ -147,13 +142,19 @@ exports.generatePDF = functions.region(index_1.REGION).https.onCall(async (paylo
         /**
          * Send whatsapp notification only if whatsapp exist
          */
-        if (((_a = data.whatsapp) === null || _a === void 0 ? void 0 : _a.template) && ((_b = data.whatsapp) === null || _b === void 0 ? void 0 : _b.recipient)) {
+        if (data.metadata.sendByWhatsapp && ((_a = data.whatsapp) === null || _a === void 0 ? void 0 : _a.template) && ((_b = data.whatsapp) === null || _b === void 0 ? void 0 : _b.recipient)) {
             await sendWhatsappNotification(data, url[0], requestedUser.business);
+        }
+        else {
+            console.log("Wont send whatsapp due to missing parameters", { sendByWhatsapp: data.metadata.sendByWhatsapp, template: (_c = data.whatsapp) === null || _c === void 0 ? void 0 : _c.template, recipient: (_d = data.whatsapp) === null || _d === void 0 ? void 0 : _d.recipient });
+        }
+        // Send document by email
+        if (data.metadata.sendByEmail) {
+            await (0, email_1.sendGenericEmail)(data, url[0], requestedUser.business);
         }
         return { url: url[0] };
     }
     catch (error) {
-        logger_1.default.error(error.message);
         throw new functions.https.HttpsError('invalid-argument', error.message);
     }
 });
