--- functions/lib/package.json
@@ -1,41 +1,0 @@
-{
-    "name": "mseller-gcloud-functions",
-    "scripts": {
-        "lint": "tslint --project tsconfig.json --fix",
-        "build": "tsc",
-        "dev": "npm run format && npm run build && firebase emulators:start --only functions",
-        "serve:all": "npm run format && npm run build && firebase emulators:start",
-        "shell": "npm run build && firebase functions:shell",
-        "start": "npm run shell",
-        "deploy": "npm run build &&  firebase deploy --only functions",
-        "logs": "firebase functions:log",
-        "tslint-check": "tslint-config-prettier-check ./tslint.json",
-        "format": "prettier --config .prettierrc.js 'src/**/*.ts' --write"
-    },
-    "engines": {
-        "node": "14"
-    },
-    "main": "lib/index.js",
-    "dependencies": {
-        "@types/cors": "^2.8.6",
-        "axios": "^1.1.3",
-        "cors": "^2.8.5",
-        "eslint-config-prettier": "^6.11.0",
-        "eslint-plugin-prettier": "^3.1.3",
-        "firebase-admin": "^11.0.1",
-        "firebase-functions": "^3.24.0",
-        "node-mailjet": "^5.1.1",
-        "pdf-documents": "^1.3.0",
-        "prettier": "^2.0.5",
-        "uuid": "^9.0.0",
-        "winston": "^3.8.2"
-    },
-    "devDependencies": {
-        "@types/uuid": "^8.3.4",
-        "firebase-functions-test": "^0.2.0",
-        "tslint": "^5.12.0",
-        "tslint-config-prettier": "^1.18.0",
-        "typescript": "^4.8.3"
-    },
-    "private": true
-}
