--- functions/lib/utls/getCurrentUser.js
@@ -1,2 +1,0 @@
-"use strict";
-//# sourceMappingURL=getCurrentUser.js.map
