+++ functions/package.json
@@ -1,5 +1,7 @@
 {
   "name": "mseller-gcloud-functions",
+  "version": "1.2.0",
+  "description": "mseller functions",
   "scripts": {
     "lint": "tslint --project tsconfig.json --fix",
     "build": "tsc",
@@ -8,6 +10,7 @@
     "shell": "npm run build && firebase functions:shell",
     "start": "npm run shell",
     "deploy": "npm run build &&  firebase deploy --only functions",
+    "deploy:function": "npm run build &&  firebase deploy --only functions:generatePDF",
     "logs": "firebase functions:log",
     "tslint-check": "tslint-config-prettier-check ./tslint.json",
     "format": "prettier --config .prettierrc.js 'src/**/*.ts' --write"
@@ -23,7 +26,7 @@
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.3",
     "firebase-admin": "^11.0.1",
-    "firebase-functions": "^3.24.0",
+    "firebase-functions": "^4.1.0",
     "node-mailjet": "^5.1.1",
     "pdf-documents": "^1.3.0",
     "prettier": "^2.0.5",
